[typ] Add explict ResultSink disable flag
Adds the --disable-resultsink flag to typ, which explicitly disables
ResultSink integration instead of relying on the presence/contents of
the LUCI_CONTEXT environment variable. The intended use case for this is
to disable ResultSink uploading if a typ test is run as part of another
test, in which case we don't want to report the "inner" results.

Also drive-by fixes an issue where we were checking
connection.getresponse() instead of connection.getresponse().status for
determining whether the data was uploaded successfully.

Bug: chromium:1143021
Change-Id: If5e72c37bfab0c4d6415cab80262c5736a0fb3ef