Add a way to skip tests
We will be disabling some tests because WebComponents V0 is being
removed from Chrome, and some tests do not work under the polyfill.
Disabling tests is useful instead of outright deleting them, because we
do not know up front if we're breaking some critical functionality that
we have to fix. If we find such critical functionality that needs to be
fixed under polyfill, we can do the fix and re-enable the functionality.

Also added examples of two test files using this skip features. These
tests are failing under the polyfill.

Bug: chromium:1111537
Change-Id: Iea1c4732f2c9303a246865825f6f34feb2282de2