Start using polyfill for tracing ui tests
WebComponents V0 is going away completely in M88, so it's time to
turn off the flags we used to keep it turned on. This CL removes
the flag for tracing UI tests.

For a lot of these cases, it's only the test case that's broken (and I
could find no easy fix.) If there is some critical functionality that's
broken under the polyfill, we can consider re-enabling the tests and
fixing the feature.

Dashboard tests still seem to require the flag, and I didn't see
any immediate fix for it, so I kept flags for dashboard for now.

Bug: chromium:1111537
Change-Id: Ibb5b78d90d0d083d4434731beff610801bc7b64f