[typ] Use requests for ResultSink
There are known slowness issues with ResultSink in various test suites
on different platforms if POSTing using a library that doesn't support
connection pooling. A typ-enabled test (telemetry_unittests) hit such an
issue on Mac 10.14, so switch to using the requests module instead of
the built-in httplib since the former supports connection pooling while
the latter does not.

As a side-effect, merges dashboard/'s .vpython file into Catapult's, as
the dashboard tests end up importing typ code, which ends up causing
issues if the requests module is not made available.

Bug: chromium:1147992
Change-Id: If0444ce06e155a0c5b0e321fcc39099aa896ce39