Add protobuf-py2_py3 3.13.0 to .vpython.
Starting from [1], //third_party/protobuf has been updated to 3.13.0
and this introduces a breaking change which breaks WebRTC's perf results
upload (see bugs.webrtc.org/12211).

I tried to add this to WebRTC's .vpython but that was not enough.
Since the proto is [2], I wonder if Catapult's .vpython should be
updated.

[1] - https://chromium-review.googlesource.com/c/chromium/src/+/2545204
[2] - https://source.chromium.org/chromium/chromium/src/+/master:third_party/catapult/tracing/tracing/proto/BUILD.gn;l=7-13;drc=c8ebd366bc298f77563a494be710b3275bc898b1

Bug: webrtc:12211
Change-Id: Ib5a8c92de0de0db8402778cfe73052714b5c7883