heap: Update gc metric names to latest naming scheme.
This CL creates a duplicate of existing blink gc metrics using the new
naming scheme. New names should provide as with metric continuity when
switching to the new library. We keep the old names for now as it would
take some time to set up alerts for all the new metrics.

Bug: chromium:1056170
Change-Id: I4e84c4265be1c887faaa9bacd9a77d7e5bab05b1