Fix NaN that appears in RunningStatistics::meanlogs
When value == 0, std::log(std::abs(value)) = -inf. It is ok. Hovewer, when the second 0 is added we end up with (-inf) - (-inf) which is NaN. A histogram that is build in such way will fail to be uploaded, since the dashboard can't handle NaN in json (since it uses ijson library).

Bug: webrtc:12224
Change-Id: If4463c2f3543a65c4aa2e2770734ca6c8b556496