chromeperf: Add sample_values to bq_rows export.
This extends the bq_rows Beam script to also export the sample values
for each Row.  Because the sample values are recorded in a separate
entity (Histogram) this complicates the Beam pipeline somewhat.  In
addition to fetching Rows for the given time period, we need to fetch
Histograms from that same period, and join these collections together.
We also need to add logic to unpack the sample values out of the
zlib-compressed JSON.

Also, now that BigQuery supports loading into clustered partitions from
clustered tables, we can restore clustering (on master, bot, measurment)
and revert the manual clustering (splitting the results into separate
tables for each master).

Change-Id: Ic2f6c7c3de905a7ee5bb7fc53b4eefc96a4984d8