Implement the unified V8 and CppGC metric computation
See bit.ly/v8-gc-stats-collection for the background and motivation.
The central part of the implementation are rules that specify which
events contribute to which metrics (see RULES in the code). This
decouples the metric computation code from event names and event layout.

New metrics can be added by adjusting the METRICS and RULES lists
without touching the metric computation code.

The new metrics are not enabled yet because we need to first land a CL
on the V8 side that adds the epoch argument to V8 events.

Bug: chromium:1154636

Change-Id: I9169e9a6ab47d4e9bb729d464697ed33d2ec9ee1