[tracing] Switch to "user_action_samples_legacy" in record dialog.
Previously, the "Chrome developer (overall)" and
"Chrome developer (navigation)" presets included both the "ui"
and "user_action_samples" categories. This resulted in duplicate
events for user actions.

https://crrev.com/c/2586106 moved the legacy user action event
event to the "user_action_samples_legacy" category. This allows
server-side config to record the "ui" category and only one
of "user_action_samples" or "user_action_samples_legacy" categories,
depending on whether the old of the new pipeline is used.

This CL updates the trace record dialog to work well with the new
categories.

Bug: chromium:1157487
Change-Id: I0d1a1769b1e8b84417bc6edf4bee4fb8d2f54f42