Upgrade mock version to 2.0.0
Upgrades the version of mock used by Catapult from 1.0.1 to 2.0.0.

As a result, adds the funcsigs and pbr modules, which are now
required for mock. This is functionally equivalent to the earlier
attempts to switch to use the version provided by vpython, but avoiding
vpython due to CrOS using Catapult but not vpython.

Bug: chromium:1156832
Change-Id: I5eaa95056e2a73f4aa86681aa8c3831e18bed022