chromeperf: Support legacy test names
An inadvertent renaming of the keys for the synthesized test names
(those that have suffixes for stats) meant we greatly reduced the
amount of anomaly detection we were doing. While we should stop storing
those legacy stats, we should still support the legacy paths until we're
ready to support the stats defined in the subscription when performing
anomaly detection.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1158326
Change-Id: Ieed35b0b54990d4bb47f601eff206e0e6292a6ad