devil: Rework the code that ensures the root privilege after reboot.
The code used to achieve this by calling EnableRoot() when
DeviceUtils.RunShellCommand(as_root=True). But this does more than
necessary and causes issues like crbug.com/1150237 and crbug.com/1152653

In the new logic, if the root privilege is granted already before the reboot,
it will be re-granted after the reboot.

Bug: chromium:1150237
Change-Id: Ia8c8a30264aef10254e333378fe7b3e4c54d80d5