Catch exceptions in _StoreUiDumpAsArtifact, watch for uiautomator output when checking correctness
ScreenDump() function can throw exceptions as uiautomator dump command
which is used inside is inherently unstable.  _StoreUiDumpAsArtifact is
itself called from exception handler, so generating new exception will
mask original one and make it harder to investigate original error.

Sadly, return code of uiautomator dump command is always 0, and in case
of an error only standard error stream may be used to check for error
condition. In such a case it will contain the string "ERROR:" as
a prefix for the message. Check this in _GetRootUiNode instead of return
code to decide whether the call was successful.

Bug: chromium:1175032
Change-Id: I025ecde2eb1963e73db9519831e3236619aa5412