Fix _PushChangedFilesZipped for the "Argument list too long" error.
When compile the gn arg "is_debug=false", the varaible "dirs" may have a
long value and it could cause the failure in device shell with the error
"Argument list too long". This CL fix this by writing the "dirs" value
into a temp file and having the script reading from it.

Bug: chromium:1174331
Change-Id: I954672006640ae2bec47da6f90c8ff63eb491b2b