Add url to new tab API
There are 2 reasons for this change:
1. Using tab.Navigate(url) requires another round trip to the browser,
and there's check to wait until navigation is complete. In comparison
tabs.New(url) is much faster.
2. Using tab.Navigate(url) to a file location will throw an error
because the navigation is never complete, while tabs.New(url) works just
fine. We recently have a use case to measure download bar performance
that requires this change to make it work.

Change-Id: Iff7f375a628a04fa1416e1f5ae759d3ca952589d