pinpoint: Enhance job listing API
This includes two enhancements:

* Fixup of the email address to hide the service account email for
  Chromeperf automation.

* Introduction of a filtering mechanism to the /jobs API.

This makes the Pinpoint UI's jobs page more deterministic and not
dependent on whether a user is signed in to see a list of jobs.

Future changes will allow us to change the UI and other consumers of the
API to filter jobs at a more granular level.

R=abennetts@google.com, fancl@google.com

Bug: chromium:1059667, chromium:1148633
Change-Id: I7a2568e4a5a7e48e6d10f27edf3f9fdebda158bb