pinpoint: Support filtering by configuration
This change supports filtering jobs by the configuration name. We also
implement support for a rudimentary and permissive filter expression
parser for query filtering. This should enable us to support the gRPC
service job listing implementation.

Also in this change, we obfuscate the email addresses that are being
shown in the jobs table.

R=abennetts@google.com, fancl@chromium.com

Bug: chromium:1059667
Change-Id: I5edda026a498b0e1576e2b3e6d018eb6f6be5a6b