trace-viewer: Don't let optional features block import
Currently if auditor and user model building (which determines various
rail stages) fail, the whole import fails. Not everyone cares about
these features, and we came across a case where an auditor was failing
because a process had two compositor threads. This CL unblocks the
import and just emits a console error.

Change-Id: Ic9ebe4198902ce9767394a3f3baf6ab1e7373be3