devil: Restart the adb server after adb commands timeout
This CL makes changes so that it restarts the adb server after
an adb command times out, which was suggested as a work around in
crbug.com/1056214. Try builders running with this change show
that retries of the adb command are able to run successfully.

Test: Created test catapult roll in crrev.com/c/2787579. Try builders
run for the roll shows that restarting the adb server after a timeout
allowed subsequent re-runs of the command to run successfully.

Bug:chromium:1056214
Change-Id: Ib91c372d8278db651f62f617abe8993fcc5263f8