[Telemetry] Add __repr__ to Error
There is at least one case in Telemetry where repr() is used on an Error
instead of just printing it out as a string. This results in all the
useful debug information included in the Error being discarded, which
defeats the entire purpose of the custom exception type. This adds a
__repr__ implementation to the class which is identical to the __str__
implementation.

Bug: chromium:1166310
Change-Id: I6457ffbf456dc019b6a406d91e0a0f52627ba1db