devil: Re-add system app modification retries and timeouts
Previously CL crrev.com/c/2091638 was merged. It removed
retries for system app modification commands that timeout. It
was merged to narrow down the root cause of timeouts that happend
when running ADB commands. Recently crrev.com/c/2787707 was merged
which implements a work around for adb commands that timeout. This
CL will re-add the retries so that the test harness execution can
recover after an adb command timeout.

Bug:chromium:1056214, chromium:1189285
Change-Id: I3734ed9a43165eccb4b5fb334ae8eef106acef68