Work around an issue in the telemetry testing framework on Mac.
If you are using a vanilla Python2.7 build (like the one
vpython will use by default now) and you run tests using
telemetry+typ, by default on the mac we'll run tests using
fork()s in the multiprocessing module.

Apparently if your tests attempt to talk directly to various
Mac frameworks (e.g., the ColorSync Framework, like
//content/test/gpu/gpu_tests/color_profile_manager_mac.py does)
using the PyObjc bridge, then the Frameworks will die and typ gets wedged.

As a grotesque workaround for this, this CL modifies the
telemetry test runner to directly import PyObjc files prior to
starting to run typ, so that the objects are copied across the
fork() safely.

Alternatives would be to not run the tests in multiple processes,
or to not use the PyObjc bridge in-process and to call out to
a script to get the needed info.

I suspect the system Python2 that is shipped on the Mac by default does
something similar implicitly to load the modules and hence we
don't see this issue.

Apparently Python3 has a "spawn" method in multiprocessing that may
avoid this, so moving to Python3 might also be an option.

Bug: chromium:1195465
Change-Id: I57686efc32cb8da9a14127f6f315c218bf006dd4