chromeperf: Create more Anomaly entities
This change makes one Anomaly entity per subscription, which alows us to
apply a different anomaly config per subscription. We change the list of
subscriptions associated with an Anomaly to contain only the matching
subscription.

This means in auto-triage, we'll create distinct Anomaly entities that
will get grouped in different alert groups by subscription. We still
have logical grouping support so that's at least a different layer of
grouping.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1158326
Change-Id: Icbe29380adfdf917df3755a33a327ce54c3fd914