devil: Add sanity check for root privilege on a device
_TearDownSystemAppModification requires root privilege on a device.
Normally root privilige is enabled on the device before calling
_TearDownSystemAppModification. The function is also re-entrant.
The function may be re-entered after the the device loses root
privilege. For instance if the adb server is restarted before
re-entering the function then the device may lose root privilege.
Therefore we need to do a sanity check for root privilege
on the device and then re-enable root privilege if the device
does not have it.

Bug:chromium:1056214
Change-Id: Id37edfd331eb64f4745c804709f59ced1050953a