systrace: check if tracefs in mounted in /sys/kernel/tracing
Currently, we only check for the existence of /sys/kernel/tracing before
returning it as the tracefs mount point. This patch parses the mount
command output to confirm tracefs in mounted on /sys/kernel/tracing
before returning it as the tracefs root.

Test: systrace with tracefs mounted in /sys/kernel/tracing, systrace
with tracefs mounted in /sys/kernel/debug/tracing.

Change-Id: I5ab53b338e786807460cebe1a69a775910c40c22