Add dispatchKeyEvent API to UIDevtools
This CL is part of the effort to support interact with native UI
for Desktop UI benchmark on the telemetry side.

Also add the Esc key and a method to key_event.py so the new API can
use them for convenience.

Doc:
https://docs.google.com/document/d/1ui3QM3tgtDL4FvpEto-vQc-LLduQH40W-eBqbttBxmI

Related CL on the browser side:
https://chromium-review.googlesource.com/c/chromium/src/+/2829258

Testing CL on Desktop UI benchmark:
https://chromium-review.googlesource.com/c/chromium/src/+/2646447

Bug: chromium:1201243
Change-Id: I47f9cb544344670bb1950f3234898f99b1d68979