[telemetry] python3 conversion: import
Convert telemetry to Python 3 style applying fixers on telemetry/:
  python -m modernize -w -f import .

This CL adds "from __future__ import absolute_import" at the beginning of each file. The reasoning can be found here:
	https://docs.python.org/3/howto/pyporting.html#prevent-compatibility-regressions

This is one of the CLs converting Telemetry to be compatible for Python3. The conversion utilizes the modernize module: https://pypi.org/project/modernize/

Bug: chromium:1198408
Change-Id: Ic5f64a4f99d13749a1dae4fe44ea537687de984d