[telemetry] python3 conversion: open
Convert telemetry to Python 3 style applying fixers on open.

This is one of the CLs converting Telemetry to be compatible for Python3. The conversion utilizes the modernize module: https://pypi.org/project/modernize/

We are seeing pylint warning like:
W:  9, 0: Redefining built-in 'open' (redefined-builtin)

But it looks like the correct approach:
https://python-future.org/compatible_idioms.html#file-io-with-open

Bug: chromium:1198408
Change-Id: I44f9c200f3e5c927dc265745e52807eee80f6c90