pinpoint: Add support for batch identifiers
This change allows Pinpoint to associate multiple jobs that are related
to the same batch. To achieve this, in this change we have:

- A top-level property for a Job called `batch_id`.

- An index to allow retrieving jobs that are related with a batch id.

- A batch_id field in the JSON retrieved from the service.

In the job listing API, we also ensure that listing the jobs in a batch
will return all the jobs in a batch. We also make job filters mutually
exclusive with batch id's for now to reduce the proliferation of indices
on the Job entity.

In the future it should be possible to apply job filters as a
post-filter on batch id queries, but for now we're limiting the support
surface to what's immediately necessary.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1203518
Change-Id: I8d97d401f03460d1628b827ac2518e1a24971865