Custom Metric: Ability to specify trace events written to the trace file
Previously, custom metric looks for trace events that start with
"custom_metric:" to show up on the result page, which does not work well
with existing trace events that doesn't start with "custom_metric:".

Instead of hardcoding the prefix,this CL reads a list of reported trace
events from a performance.mark() from the trace file to select those
events. It allows any stories from telemetry to specify any trace events
of interest to show up on the result page.

Bug: chromium:1201243
Change-Id: I1491aeea8a63a72a02ab652a70c45d244d7461c4