Move code in systrace/bin/adb_profile_chrome_startup into .py script.
We can call an android trace from functions in scripts located
outside of catapult by moving the current code out of a binary
and into a .py script. This function (ProfileChrome) is useful
in building a tracing pipeline because it finds where the trace
file is written and automatically copies it to a local folder.

Bug: 1217859

Change-Id: Ib2d6a6f810dd86ddf145f7bbbbdf1a726cb549b6