[typ] Decode b64 data
Decodes the output from base64.b64encode to UTF-8 when adding the
stdout/stderr as artifacts in result_sink.py, as the returned bytes
cause issues with JSON encoding in Python 3.

There are other Python 3 issues in typ still, but this appears to be the
minimum change to get the grit tests passing and unblock a user - the
other issues will be handled in another change.

Bug: chromium:1222840
Change-Id: Ib73dd2aef0cb6a5ed08965a9baf94bab3e902e88