[Telemetry] Remove finalizer from browser backends
Some browser backends have a finalizer (__del__ method). This is very
problematic, as Python makes no guarantee in the order of calling
finalizers. In particular, other objects used by the backends to keep
track of Chrome process and DevTools connection may have already been
finalized. This is causing hard-to-debug race conditions when running
with Python 3.

To avoid the race conditions, this CL removes the finalizers. The code
using the backends should remember to call Close at the end. (If any
code forgets to Close the backend, by the time finalizer is called,
it is too late to reliably clean up.)

Bug: chromium:1198408
Change-Id: I2f772bc9621d1e4928cc6f379cde45aced51c7f9