Fix flakiness in test `ActualPageRunEndToEndTests.testTrafficSettings`.
This test checks that slow page is slower than fast page by at least
40 - 2 = 38ms. In some environments this condition is wrong.
Technically, if we had an ideal system timer and other processes do not
affect time measurement, this check looks reasonable. But in the real
world system timer may have granularity greater than 1ms, and time
measurements are more affected by another processes when the fast page
loading and less when the slow page is loading. In our testing
environment even with fast developer machines latency time of the fast
page is very often greater than latency time of the slow page. Thus we
need change testing traffic settings for this test in order to run is
stably. Fast page must be really fast and slow must be really slow.
We changed this settings by increasing round-trip latency and
decreasing bandwidth of slow page and vice versa for fast page.
This change improves stability of this test in our environment.

Change-Id: I8b5218430270fa531b5eea174e08b0cc04892f95