Fix LCP invalidate event filtering
The code is currently filtering using isLcpInvalidateEvent, which is
meant to be used on a regular trace event slice event, not the custom
LcpEvent class defined in this class. The LcpEvent class doesn't have
the title property, so this check will always return false and the
invalidate events are not filtered out.

The invalidate event doesn't have the "durationInMilliseconds" field
so the sample values become undefined later, so the LCP is still
reported as empty. The bug therefore doesn't have much impact currently
beyond confusing developers stepping through the code.

Bug: chromium:1227603
Change-Id: I146dd86d32f7bfcf2305d4b458ec3ce9afc9c0de