[catapult] unblock Build Python Tests on Catapult tryjob
This change help unblock the running of dashboard/bin/run_py_test on my local linux machine. However, the PY3 try job is still failing. The error message showed import error on:
File "/b/s/w/ir/k/catapult/catapult_build/dev_server.py", line 11, in <module>
  import urlparse

The specific line should be updated by this CL but somehow the change is not taken in the test. I'll try to check this in to see how the change goes in the next CL.

Bug: chromium:1221663
Change-Id: Idc868f0c560fedce06bfc81d156e7b18f8891db3