Android: Add support for testing play core
Chromium is updating play core and will be using its new --local-testing
support. This requires that the fake module apks be copied to a new
location under the apk's data directory. This CL supports both the old
location and the new location. Once Chromium completes its migration to
the new play core library, support for the old location can be removed.

Bug: chromium:1220662, chromium:1000751
Change-Id: I066e486cf67ad2fcbef50abefb972bc9a036b344