Android: Add support for testing play core (reland)
Chromium is updating play core and will be using its new --local-testing
support. This requires that the fake module apks be copied to a new
location under the apk's data directory. This CL supports both the old
location and the new location. Once Chromium completes its migration to
the new play core library, support for the old location can be removed.

Previously reverted for android-11 bot failure, but the issue is with
the bot's emulator image instead of this CL.

Original CL: https://crrev.com/c/3024597

Bug: chromium:1220662, chromium:1000751, chromium:1234375