[Pinpoint] Make fallback targets lists
Makes the logic related to fallback targets use lists instead of
strings so multiple fallback targets can be specified if needed. Strings
are still supported, as they will be automatically converted to a list.

This is so that Pinpoint will (hopefully) continue to function properly
with some potential changes to which isolates are used on certain bots.

Bug: chromium:1231290
Change-Id: I0a43953817daada7a87d5ad5bd7e7b0221d460c5