Fix testScreenshot and enable this test in Windows
When launching chromium without explicitly set color scheme, chromium
uses Default scheme, which is not sRGB. When testScreenshot checks
color of pixel, it checks pixel in RGB, but screenshot was taken from
chromium, launched with NOT sRGB color scheme. This cause test failure
on Linux and Windows. This CL explicitly sets color scheme to sRGB
only for testScreenshot and enables test on Windows.

Change-Id: I55768a26c8253a46d43ed2857075c09c18f12f2c