[typ] Add ability to force retry on failure
Adds the ability for a test to signal to typ that the current test
should be retried on failure instead of only relying on expectation
files. The current use case for this is a temporary workaround for the
WebGL conformance tests flakily failing the first test on a shard,
although there may be other cases where a test would like to signal
that it should be retried.

Bug: chromium:1271227
Change-Id: I5b0dae145736fb7378e798f82f04f48141a074f7