[Catapult] Return binary strings on stop tracing
When Python 3 is enabled on perf waterfall, the story system_health.webview_startup/load:chrome:blank started to fail on webview and weblayer builders:
https://ci.chromium.org/ui/p/chrome/builders/ci/Android%20Nexus5X%20WebView%20Perf/19377/test-results

On the Clank builders, it looks fine, such as shard#7 in:
https://ci.chromium.org/ui/p/chrome/builders/ci/android-pixel2-perf/21075/overview
https://cas-viewer.appspot.com/projects/chrome-swarming/instances/default_instance/blobs/eeb81fed316e7ee7e595230cd7fefcf3c225e595c6ba4d5c184ce55c046805aa/1184141?filename=benchmark_log.txt

The failure happens when _stop_collect_trace() returns a string instead of binary on the non-clank (webview/weblayer) setup. I'm using the ensure_binary from six here to enforce the same behavior between clank and non-clank.

bug: chromium:1271668
Change-Id: I70eee62ae0e955b4e49e66d8022fe65412869058