[Pinpoint] Run single test in python 2 for old commits
This is a workaround to run bisect on old commits. When the commit position is older than an anchor commit X, the commit will be run in python 2.

Please notice that this is not the optimal solution. A complete fix should allow all tests within a bisect to run on python 2, when the anchor commit X is between the start and end commit, but the chance is rare. We can add an argument to run bisect/tryjob in python 2 if needed in a future CL.

Bug: chromium:1278382
Change-Id: I6bcf247520be4f376379e52e328e3bee9e6bd406