Short summary
This cl is intended improve two things.
Firstly, it is intended to pass line and filepath of
failed test to trace to simplify navigation through failed
tests and make trace more informative for CI.
Secondly, it is intended to use 'real_test_func'
attribute, if it is specified, to get real path of
generated tests. This change is made to use with
Chromium benchmarks' unittests: shorturl.at/bjsP1
Right now benchmarks' unittests would write test generator
filepath to trace, but using 'real_test_func' would make it
easy to forward benchmark class to typ and write line of
code and fileapth of failed benchmark instead.

Change-Id: I0bf87a235af1389a438e5093f16c27bde59c6539