[inspector] Disable DevTools call stack capturing.
In crbug/1280831 it was found that a seamingly innocent debugger-only
refactoring (crrev.com/c/3302794) had a surprisingly significant impact
on various system health benchmarks. Further investigation revealed that
catapult is using the Chrome DevTools Protocol, and in particular it's
Runtime domain, to drive these tests. When enabling the Runtime domain
however, V8 starts collecting detailed stack traces (for the inspector)
in addition to the simple stack traces (for Error.stack), and the recent
refactoring (crrev.com/c/3302794) made the collection of detailed stack
traces slightly slower by moving some of the overhead from the inspector
to the point where the stack trace is captured (motivated by the
eventual ability to manage stack traces in the V8 heap and thus
drastically cut costs for the debugger both in terms of performance and
in terms of memory).

In reality the refactoring (crrev.com/c/3302794) does not have any
noticable performance impact, since end users don't use any DevTools
functionality, so the regressions observed in crbug/1280831 are purely
an artifact of the test setup and framework.

With crrev.com/c/1085643, the inspector already received the ability to
control the stack capturing via Runtime.setMaxCallStackSizeToCapture, so
we should explicitly disable the capturing by setting this flag to 0 in
the test setup, which will not only recover the performance regression
mentioned above, but should also make the measured performance match the
actual (real world) performance more closely, since V8 will stop to
capture stack traces for the inspector completely.

Doc: https://bit.ly/v8-stack-frame
Bug: chromium:1278647, chromium:1258599, chromium:1278650
Change-Id: Ifd4131bf38d714aa434490f0cbfbd3cd4b019600
Fixed: chromium:1280831