[typ] Add skipTest support
Adds support support to typ for cases where a test calls skipTest().
Under the built-in Python unittest runner, this causes the test to be
skipped, but under typ, it currently just causes an exception to be
raised.

Also makes some minor drive-by fixes to unittests that were found
while writing the unittest for the new behavior.

Change-Id: I17b4a812f7aac0bd11397c78feedc2d62860ca34