[typ] Add skipTest support
Adds support support to typ for cases where a test calls skipTest().
Without this change, any uses of skipTest() get treated as unexpected
skips by typ. With this change, tests can set a boolean in TestCase
signalling that the skip is in fact expected and should be handled
accordingly.

Also makes some minor drive-by fixes to unittests that were found
while writing the unittest for the new behavior.

Bug: 1285055

Change-Id: I17b4a812f7aac0bd11397c78feedc2d62860ca34