[Sheriff] Add log info to testSystemInfoModelNameOnMac
Test has been failing on bots, but it's not clear why it's failing
and what CL caused the failure. The failure happens in the line where
it check if the info_model satisfies the re pattern, but there is no
info about what is being matched against in logs. Adding some logging
to help us identify what's up here.

Bug: chromium:1286736
Change-Id: I736e4116ad22c62c0d98f04b2360a3ee499f93a7