Add luci_proto third_party library.
This contains all the protos necessary to interact with LUCI services.
Our immediate need is just the protos for the Buildbucket v2 API, but
it's easier to import the entire tree wholesale.

Includes easy update script which records the last-imported commit.

I'm piggy-backing off of the infra.git generate proto files because
correctly generating them from the source of truth (the luci-go repo)
is extra-painful (and would result in the same output anyway).

R=fmmirzaei@google.com, seanmccullough@google.com

Bug: chromium:1264738
Change-Id: Iad0a4068855a6d70d558b1376c9cf756e33cdfc8