Use bbagent in the recipe definitions.
This will cause the generated buildbucket config to use the exe and
properties fields instead of the recipe field. This is implied when
setting use_python3 = true, so this will keep the config changes for
that isolated. Builders are now all switched to using bbagent by the
buildbucket service, so this should cause no functional changes.

Bug: chromium:1256437
Change-Id: Id94d5f1c2e298aa5abb6feb861f27295d80de61d