Clarify stale file behavior and variable naming.
This adds a test that shows that pushing
 a/b/c.txt
results in other files in a/b/ considered stale so e.g.
 a/b/d.txt
 a/b/e/...
get nuked.

Variable naming was a bit confusing so I tried to improve it.
"dir" is now referring to a directory and "path" can be directory
or file. Initially host_device_tuples can be a mix of both.

Bug: angleproject:6854
Change-Id: Iae7d410ebba149701c331e528cf6af944621d078