Fix to InstallMultiple swallowing the first arg
Currently `adb install-multiple` ignores the first arg that is passed
through to package manager. This is being fixed but in the meantime, add
this workaround to avoid missing args causing `adb install-multiple` to
fail.

This is somewhat less of a problem on Android API 28+ since oftentimes
`-r` is the first arg passed through and on API 28+ `-r` is assumed by
default, so it being missing did not cause `adb install-multiple` to
behave incorrectly on those API versions. However, this is very brittle
so an explicit workaround is better for all API versions and in case
the first arg is not -r.

Bug: b/218716282
Change-Id: Ia0f9c5da55e601ed65c20240b124814c782f8a4b