Teach fuchsia_interface about the Mac
On the Mac, `platform.machine` returns "arm64", not "aarch64". Make
this response valid.

FWIW, it's not clear to me if this is the correct thing to do. This
file says that it's a "wrapper for common operations on a Fuchsia
device" which implies I shouldn't be making changes to it to run on
non-Fuchsia devices. On the other hand, this script _is_ running on
non-Fuchsia devices, so I'm going to make this change.

Bug: chromium:1294584
Change-Id: I15703ba4cfd752d3a0f2c10c3ac83eb966cd132e