metrics: Report CLS value as 0 if no LayoutShift event found
We were reporting null instead of 0 if no LayoutShift event was found.
This fixes the problem for single navigation cases which is the vast
majority of our tests.

This doesn't quite work for multiple navigation case; ideally we should
be emitting a separate CLS value for each navigation. But multiple
navigation wasn't handled properly before anyways, so this CL doesn't
make it any worse.

Bug: chromium:1290843
Change-Id: Ibb376d6bf9c7e6d26875e405b33858704775b0fb