[fuchsia] Fixes telemetry on Fuchsia-Chrome to run multiple stories.
This CL will fix telemetry on Fuchsia-Chrome so that it can run:
- a single story with ReturnAsStream tracing
- stories that use WPR (most benchmarks)
- multiple stories at once (as most benchmarks are run).

This was done via the following changes:
- forward the WPR args to the browser. These args are necessary
to indicate to the browser the proxy server it should use during
the tests.
- kill the browser correctly. Before this change, the browser was
killed via context_provider.cmx, which is not used by fuchsia-chrome
- fetch the logs in a fashion that does not conflate previous browser
instances, and enable logging (--enable-logging and --v=3). Verbose
logging is required to get the DevTools port printed, and logging
must be examined only after the instance is created, not from before.
In the previous case, DevTools log messages from previous browser
instances were being picked up, causing port forwarding to be launched
to old DevTools ports (thus failing).

With this change, ~50 system_health.common_desktop benchmarks could
run successfully on fuchsia-chrome.

Change-Id: I2db1d4c772f195a4803dc4a4b1f1ff13c618d9fc
Bug: 1297360